style: clear the clippy backlog and gate it in CI + the hook
`[workspace.lints.rust] warnings = "deny"` covers clippy lints too, so
the ~15 outstanding lints meant `cargo clippy` was broken for everyone in
this repo, not just unenforced. Cleared, then gated so it can't rot
again. No behaviour change — 529 tests pass, all three backends.

Two are #[allow] with a reason rather than clippy's rewrite:

  * mob_update.rs keeps `(800u64 << 0) | (16u64 << 19) | (2400u64 << 45)`
    under identity_op. The `<< 0` is the first field's bit POSITION,
    parallel to 19 and 45, and the asserts below cite exactly those
    ranges — folding it to `800u64` hides where the pack starts.
  * the #[cxx::bridge] module gets incompatible_msrv: cxx expands
    `-> Box<T>` into an API newer than the declared rust-version 1.75.
    It's generated, so there is nothing here to rewrite.

player_self_pos.rs's precedence lint took parens instead — that keeps the
fixture byte-identical and reads better, so an allow wasn't warranted.

One doc lint was a real bug, not style: a `///` block describing
OP_ClientUpdate sat above parse_new_zone separated only by `//` lines, so
rustdoc was attaching an unrelated wire layout to that function. It's `//`
now. The rest are wire-layout comments that markdown was parsing as lists
— parse_buff_list's diagram moved into a ```text fence, matching
guild_expanded_info.rs.

Clippy runs per backend in both CI and the hook: the feature-gated FFI
arms in seq-bridge are only linted under the feature that enables them.
Verified the gate bites — the hook exits 101 on a clippy-only violation
that still compiles and tests green.

fmt stays ungated (dirty in 408 places); reformatting would swamp git
blame across the wire-format archaeology. Noted in CLAUDE.md.

+++ seq-bridge/src/lib.rs
@@ -28,6 +28,10 @@ use seq_decode as backend;
 #[cfg(feature = "backend-eql")]
 use seq_backend_eql as backend;
 
+// cxx expands `-> Box<T>` entries into code using an API stabilized after the
+// workspace's declared rust-version (1.75). It's generated, so there is nothing
+// here to rewrite; the toolchain we actually build with is `stable`.
+#[allow(clippy::incompatible_msrv)]
 #[cxx::bridge(namespace = "seq::rust")]
 mod ffi {
     /// Decoded `OP_MobUpdate` payload. `ok` is the discriminator: when
@@ -1365,8 +1369,8 @@ fn decode_mob_health(bytes: &[u8]) -> ffi::MobHealth {
 fn decode_spawn_appearance(bytes: &[u8]) -> ffi::SpawnAppearance {
     match backend::parse_spawn_appearance(bytes) {
         Ok(a) => ffi::SpawnAppearance {
-            spawn_id: u32::from(a.spawn_id),
-            kind: u32::from(a.kind),
+            spawn_id: a.spawn_id,
+            kind: a.kind,
             // Live's current wire carries no value field; eql's still does.
             #[cfg(feature = "backend-eql")]
             parameter: a.parameter,
